UtilをUIとStatsに分離
